+++ src/PrestaShopBundle/Bridge/Helper/Listing/Action/ListHeaderToolbarAction.php
@@ -26,13 +26,13 @@
 
 declare(strict_types=1);
 
-namespace PrestaShopBundle\Bridge\Listing\Action;
+namespace PrestaShopBundle\Bridge\Helper\Listing\Action;
 
 use PrestaShopBundle\Bridge\AdminController\Action\AbstractAction;
 
 /**
- * This class is the object to instantiate if you want to add a bulk action for your list.
+ * This class is the object to instantiate if you want to add an action in the header toolbar of your list.
  */
-class ListBulkAction extends AbstractAction
+class ListHeaderToolbarAction extends AbstractAction
 {
 }
