+++ src/PrestaShopBundle/Bridge/Helper/Listing/HelperListConfigurationFactory.php
@@ -26,7 +26,7 @@
 
 declare(strict_types=1);
 
-namespace PrestaShopBundle\Bridge\Listing;
+namespace PrestaShopBundle\Bridge\Helper\Listing;
 
 use PrestaShopBundle\Bridge\AdminController\ControllerConfiguration;
